Maintain ordering (giving up stability).  It seems to slow down inserts without speeding up find (even unsuccessful find).
Is this surprising?  On the one hand, in the stable version, most inserts occur at a low load factor, so successful searches should mostly be fast even without ordering.

On the other hand, unsucessful searches become more likely to have to pay.

Better instrumentatin is needed to understand this.  What is the average unsuccessful search probe length, for example?